backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Production server
gunicorn==23.0.0

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0

# Database migrations
alembic==1.14.0

# Telegram
aiogram==3.16.0

# Scheduling
apscheduler==3.10.4

# HTTP client
aiohttp==3.10.11

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
greenlet==3.1.1

# JWT auth
PyJWT==2.10.1

# Redis client
redis[hiredis]==5.2.1
